Make use of new ledger methods
