+++ bcos-scheduler/src/SchedulerImpl.h
@@ -164,9 +164,6 @@ private:
     BlockExecutive::Ptr getLatestPreparedBlock(bcos::protocol::BlockNumber blockNumber);
     void tryExecuteBlock(bcos::protocol::BlockNumber number, bcos::crypto::HashType parentHash);
 
-    void asyncGetLedgerConfig(
-        std::function<void(Error::Ptr, ledger::LedgerConfig::Ptr ledgerConfig)> callback);
-
     BlockExecutive::Ptr getPreparedBlock(
         bcos::protocol::BlockNumber blockNumber, int64_t timestamp);
