+++ bcos-scheduler/test/CMakeLists.txt
@@ -9,7 +9,7 @@ target_compile_options(${TEST_BINARY_NAME} PRIVATE -Wno-unused-variable)
 
 find_package(Boost REQUIRED unit_test_framework serialization)
 
-target_link_libraries(${TEST_BINARY_NAME} ${SCHEDULER_TARGET} ${CRYPTO_TARGET}
+target_link_libraries(${TEST_BINARY_NAME} ${SCHEDULER_TARGET} ${LEDGER_TARGET} ${CRYPTO_TARGET}
 Boost::unit_test_framework ${TARS_PROTOCOL_TARGET} Boost::serialization ${STORAGE_TARGET} ${SECURITY_TARGET})
 
 add_test(NAME test-scheduler WORKING_DIRECTORY ${CMAKE_RUNTIME_OUTPUT_DIRECTORY} COMMAND ${TEST_BINARY_NAME})
